Add beholder metrics
